+++ src/cpp11.cpp
@@ -138,6 +138,13 @@ extern "C" SEXP _capybara_sqrt_(SEXP w) {
     return cpp11::as_sexp(sqrt_(cpp11::as_cpp<cpp11::decay_t<const SEXP &>>(w)));
   END_CPP11
 }
+// 05_kendall_correlation.cpp
+double kendall_cor_(doubles_matrix<> m);
+extern "C" SEXP _capybara_kendall_cor_(SEXP m) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(kendall_cor_(cpp11::as_cpp<cpp11::decay_t<doubles_matrix<>>>(m)));
+  END_CPP11
+}
 
 extern "C" {
 static const R_CallMethodDef CallEntries[] = {
@@ -150,6 +157,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_capybara_group_sums_spectral_", (DL_FUNC) &_capybara_group_sums_spectral_, 5},
     {"_capybara_group_sums_var_",      (DL_FUNC) &_capybara_group_sums_var_,      2},
     {"_capybara_inv_",                 (DL_FUNC) &_capybara_inv_,                 1},
+    {"_capybara_kendall_cor_",         (DL_FUNC) &_capybara_kendall_cor_,         1},
     {"_capybara_rank_",                (DL_FUNC) &_capybara_rank_,                1},
     {"_capybara_sandwich_",            (DL_FUNC) &_capybara_sandwich_,            2},
     {"_capybara_solve_beta_",          (DL_FUNC) &_capybara_solve_beta_,          4},